# zap/openapi.yaml
openapi: "3.0.0"
info:
  title: Basesource API
  version: "1.0"
servers:
  - url: http://host.docker.internal:8080
paths:
  /health:
    get: {}
  /api/v1/todos:
    get:
      parameters:
        - {name: search, in: query, schema: {type: string}}
        - {name: sort_by, in: query, schema: {type: string}}
        - {name: done, in: query, schema: {type: boolean}}
        - {name: page, in: query, schema: {type: integer}}
        - {name: limit, in: query, schema: {type: integer}}
    post: {}
  /api/v1/todos/{id}:
    get:
      parameters: [{name: id, in: path, required: true, schema: {type: integer}}]
    put:
      parameters: [{name: id, in: path, required: true, schema: {type: integer}}]
    delete:
      parameters: [{name: id, in: path, required: true, schema: {type: integer}}]
  /api/v1/auth/register:
    post:
      parameters:
        - {name: sort, in: query, schema: {type: string}}
  /api/v1/auth/login:
    post:
      parameters:
        - {name: sort, in: query, schema: {type: string}}
  /api/v1/auth/refresh:
    post: {}
  /api/v1/auth/logout:
    post: {}
  /api/v1/me:
    get: {}
  /api/v1/my/todos:
    get:
      parameters:
        - {name: search, in: query, schema: {type: string}}
        - {name: sort_by, in: query, schema: {type: string}}
    post: {}
  /api/v1/my/todos:
    delete:
      parameters:
        - {name: condition, in: query, required: true, schema: {type: string}}
  /api/v1/my/todos/{id}:
    get:
      parameters: [{name: id, in: path, required: true, schema: {type: integer}}]
    put:
      parameters: [{name: id, in: path, required: true, schema: {type: integer}}]
    delete:
      parameters: [{name: id, in: path, required: true, schema: {type: integer}}]
  /api/v1/my/todos/{id}/field:
    patch:
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
        - {name: field, in: query, required: true, schema: {type: string}}
        - {name: value, in: query, schema: {type: string}}
  /api/v1/my/todos/{id}/share:
    post:
      parameters: [{name: id, in: path, required: true, schema: {type: integer}}]
  /api/v1/my/todos/{id}/share/{uid}:
    delete:
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
        - {name: uid, in: path, required: true, schema: {type: integer}}
  /api/v1/my/todos/{id}/attachment:
    post:
      parameters: [{name: id, in: path, required: true, schema: {type: integer}}]
    delete:
      parameters: [{name: id, in: path, required: true, schema: {type: integer}}]
  /api/v1/admin/todos/bulk-delete:
    post:
      parameters:
        - {name: title_filter, in: query, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items: {type: integer}
                  example: [1]
            example:
              ids: [1]
  /api/v1/admin/todos/bulk-status:
    patch:
      parameters:
        - {name: order_by, in: query, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items: {type: integer}
                  example: [1]
                done:
                  type: boolean
                  example: true
            example:
              ids: [1]
              done: true
  /api/v1/todos/{id}/comments:
    get:
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
        - {name: order_by, in: query, schema: {type: string}}
    post:
      parameters: [{name: id, in: path, required: true, schema: {type: integer}}]
  /api/v1/todos/{id}/comments/{cid}:
    delete:
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
        - {name: cid, in: path, required: true, schema: {type: integer}}